pyproject.toml
[project]
name = "codereview-env"
version = "1.0.0"
description = "OpenEnv-compatible RL environment for code review and PR management"
authors = [
    {name = "Hackathon Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.3",
    "httpx>=0.28.1",
    "openai>=1.57.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
